+++ crates/config/tests/generator_test.rs
@@ -53,6 +53,7 @@ struct GenConfig {
     id: warpgate::Id,
     locator: Option<warpgate::PluginLocator>,
     json_value: serde_json::Value,
+    toml_value: Option<toml::Value>,
     yaml_value: serde_yaml::Value,
 }