+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Added `type_serde_json`, `type_serde_toml`, and `type_serde_yaml` features, that implements
   schematic types for serde values.
 
+#### 🐞 Fixes
+
+- Updated json schema unknown/any type to be a union of all types, instead of null.
+
 #### ⚙️ Internal
 
 - Updated dependencies.